Add support for using `salt-cloud` to deploy VMs on Ryhino
This is still in its early stages. This commit mainly aims to lay down
the foundation for future flush-outs.

States defined in this commit can be used to initialize a new salt
master that can handle everything the prod can, i.e. the prod is now
saltified.

This commit adds:
- Small amount of salt-master configuration
- Config to connect to the bedrock vCenter
- GPG-based secret rendering
- One template-based VM profile with Rocky Linux directly on Obsidian
  and infra VLAN.