Pull in `salt-bootstrap` through submodule; Use dev branch
This commit addresses RH-21

At the time of writing (2023/02/02), the bootstrap-salt.sh script is
still stuck at v2022.10.04, which has some compatibility issues with
Rock Linux 9.

So, the solution in this PR is to just pull in development branch of
`salt-bootstrap` and copy the deploy script to salt masters through the
typical `file.managed` state. This way, we also get version pinning for
free, since submodule relies on commit hashes.