Add `docker-formula` to GitFS; Add compat layer
This commit adds `docker-formula` to GitFS so that we can install and
manage Docker on a target. An compatibility layer is introduced in order
to bridge some incompatibilities as described in RH-15 and also to raise
warning instead of throwing an error when the formula has not been
cloned yet.